# website/pages/assets/openapi.yaml
openapi: 3.1.0
info:
  title: PicoMQ HTTP API
  version: 0.1.0
  description: >
    The Pico protocol, PicoMQ's native HTTP API, plus the admin API served by
    each node's admin listener. A stream is addressed by its path. Requests for
    a stream owned by another node return 307 with the owner's address. With
    auth required, send `Authorization: Bearer <token>` (probes stay open).
servers:
  - url: http://127.0.0.1:4437
    description: Protocol listener
...
security:
  - BearerAuth: []
paths:
...
  /admin/cluster:
    get:
      tags: [admin]
      summary: Cluster overview
      servers: [{ url: "http://127.0.0.1:9090" }]
      responses:
        "200":
          description: Identity, counts, applied index, lease holder, pending transfers.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cluster" }
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
  /admin/nodes:
    get:
      tags: [admin]
      summary: Registered nodes
      servers: [{ url: "http://127.0.0.1:9090" }]
      responses:
        "200":
          description: Every registered node.
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items: { $ref: "#/components/schemas/Node" }
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
  /admin/streams/{stream}:
    get:
      tags: [admin]
      summary: One stream's placement and state
      servers: [{ url: "http://127.0.0.1:9090" }]
      parameters:
        - $ref: "#/components/parameters/Stream"
      responses:
        "200":
          description: Owner, state, epoch, offsets, pending transfer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamDetail" }
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "404":
          description: No such stream.
  /admin/transfer:
    post:
      tags: [admin]
      summary: Transfer a stream to another node
      servers: [{ url: "http://127.0.0.1:9090" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stream, toNode]
              properties:
                stream: { type: string }
                toNode: { type: integer }
      responses:
        "200":
          description: Transfer recorded. Completion is asynchronous.
          content:
            application/json:
              schema:
                type: object
                properties:
                  stream: { type: string }
                  streamId: { type: integer }
                  toNode: { type: integer }
                  pending: { type: boolean }
        "400":
          description: Invalid target or a conflicting transfer is pending.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
  /admin/nodes/{id}:
    post:
      tags: [admin]
      summary: Update a node's placement slots
      servers: [{ url: "http://127.0.0.1:9090" }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slots]
              properties:
                slots: { type: integer, minimum: 0 }
      responses:
        "200":
          description: The updated node.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "404":
          description: No such node.
  /admin/tokens:
    get:
      tags: [tokens]
      summary: List tokens
      servers: [{ url: "http://127.0.0.1:9090" }]
      responses:
        "200":
          description: Records visible to the caller. Secrets are never returned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenList" }
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
    post:
      tags: [tokens]
      summary: Issue a token
      servers: [{ url: "http://127.0.0.1:9090" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueTokenRequest" }
      responses:
        "201":
          description: Issued. The wire token appears once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssueTokenResponse" }
        "400":
          description: Invalid id or scope.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminError" }
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "409":
          description: Token id already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminError" }
  /admin/tokens/{id}:
    delete:
      tags: [tokens]
      summary: Revoke a token
      servers: [{ url: "http://127.0.0.1:9090" }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Revoked.
        "401":
          $ref: "#/components/responses/AdminUnauthenticated"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "404":
          description: No such token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminError" }
        "409":
          description: Token changed concurrently.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminError" }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    Stream:
      name: stream
      in: path
      required: true
      description: The stream name. Names are paths and may contain slashes.
      schema: { type: string }
...
  responses:
...
    AdminUnauthenticated:
      description: Missing or invalid credential.
      headers:
        WWW-Authenticate:
          schema: { type: string, enum: [Bearer] }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
    AdminForbidden:
      description: Valid credential, insufficient scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AdminError" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        next_seq: { type: integer }
    AdminError:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Cluster:
      type: object
      properties:
        clusterId: { type: string }
        nodeId: { type: integer }
        nodeEpoch: { type: integer }
        advertisedAddress: { type: string }
        registered: { type: boolean }
        appliedIndex: { type: integer }
        streamCount: { type: integer }
        objectCount: { type: integer }
        streamSetObjectCount: { type: integer }
        streamObjectCount: { type: integer }
        preparedObjectCount: { type: integer }
        kvEntryCount: { type: integer }
        kvBytes: { type: integer }
        gc:
          type: object
          description: Destroyed-object cleanup backlog.
          properties:
            backlog: { type: integer }
            oldestSeq: { type: [integer, "null"] }
            nextSeq: { type: integer }
        snapshot:
          type: object
          description: Metadata snapshot cycle stats.
          properties:
            lastAppliedIndex: { type: integer }
            lastBytes: { type: integer }
            lastDurationMs: { type: integer }
            lastAtMs: { type: integer }
            taken: { type: integer }
            failed: { type: integer }
        leaseHolder:
          type: [boolean, "null"]
          description: Whether this node holds the maintenance lease. Null when unknown.
        pendingTransfers:
          type: array
          items:
            type: object
            properties:
              streamId: { type: integer }
              fromNode: { type: integer }
              toNode: { type: integer }
    Node:
      type: object
      properties:
        nodeId: { type: integer }
        nodeEpoch: { type: integer }
        advertisedAddress: { type: string }
        slots: { type: integer }
        openingCount: { type: integer }
        placedCount: { type: integer }
        local: { type: boolean }
    StreamDetail:
      type: object
      properties:
        name: { type: string }
        streamId: { type: integer }
        ownerNodeId: { type: [integer, "null"] }
        ownerAdvertisedAddress: { type: [string, "null"] }
        ownerLocal: { type: boolean }
        contentType: { type: string }
        ttlSeconds: { type: [integer, "null"] }
        expiresAtMs: { type: [integer, "null"] }
        closed: { type: boolean }
        state:
          type: [string, "null"]
          description: Placement state, opening or placed. Null before first open.
        epoch: { type: [integer, "null"] }
        nodeId:
          type: [integer, "null"]
          description: Node in the placement row, may lag the effective owner during a transfer.
        startOffset: { type: [integer, "null"] }
        endOffset: { type: [integer, "null"] }
        pendingTransfer:
          type: [object, "null"]
          properties:
            fromNode: { type: integer }
            toNode: { type: integer }
    ResourceMatcher:
      type: object
      oneOf:
        - required: [exact]
          properties:
            exact: { type: string }
        - required: [prefix]
          properties:
            prefix: { type: string }
    ReadWrite:
      type: object
      properties:
        read: { type: boolean }
        write: { type: boolean }
    Scope:
      type: object
      properties:
        streams:
          type: array
          items: { $ref: "#/components/schemas/ResourceMatcher" }
        tokens:
          type: array
          items: { $ref: "#/components/schemas/ResourceMatcher" }
        groups:
          type: object
          properties:
            stream: { $ref: "#/components/schemas/ReadWrite" }
            tokens: { $ref: "#/components/schemas/ReadWrite" }
            admin: { $ref: "#/components/schemas/ReadWrite" }
        ops:
          type: array
          items:
            type: string
            enum:
              [
                read,
                head,
                list,
                create,
                append,
                trim,
                close,
                delete,
                issue_token,
                revoke_token,
                list_tokens,
                cluster_read,
                node_read,
                stream_inspect,
                transfer_stream,
                update_node_slots,
              ]
        audiences:
          type: array
          items: { type: string, enum: [pico, durable_streams, admin] }
        autoPrefixStreams: { type: boolean }
        expiresAtMs: { type: [integer, "null"] }
    TokenRecord:
      type: object
      properties:
        id: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        createdAtMs: { type: integer }
        issuedBy: { type: string }
    TokenList:
      type: object
      properties:
        count: { type: integer }
        tokens:
          type: array
          items: { $ref: "#/components/schemas/TokenRecord" }
    IssueTokenRequest:
      type: object
      required: [id, scope]
      properties:
        id: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
    IssueTokenResponse:
      type: object
      properties:
        id: { type: string }
        token: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        createdAtMs: { type: integer }